Feature/FIN-235 setup a relation between contrac

+++ jpa-repository/src/main/java/com/jongsoft/finance/jpa/account/AccountSpendingImpl.java
@@ -1,15 +1,17 @@
 package com.jongsoft.finance.jpa.account;
 
 import com.jongsoft.finance.domain.account.Account;
-import com.jongsoft.finance.domain.account.AccountProvider;
+import com.jongsoft.finance.providers.AccountProvider;
+
+import java.math.BigDecimal;
 
 public class AccountSpendingImpl implements AccountProvider.AccountSpending {
 
     private final Account account;
-    private final double total;
+    private final BigDecimal total;
     private final double average;
 
-    public AccountSpendingImpl(Account account, double total, double average) {
+    public AccountSpendingImpl(Account account, BigDecimal total, double average) {
         this.account = account;
         this.total = total;
         this.average = average;
@@ -22,7 +24,7 @@ public class AccountSpendingImpl implements AccountProvider.AccountSpending {
 
     @Override
     public double total() {
-        return total;
+        return total.doubleValue();
     }
 
     @Override
